rename: mj-sql-parser.ts → sql-parser.ts
