+++ packages/SQLParser/src/__tests__/edge-cases.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { SQLParser } from '../mj-sql-parser.js';
+import { SQLParser } from '../sql-parser.js';
 import { MJLexer } from '../mj-lexer.js';
 import { MJPlaceholderSubstitution } from '../mj-placeholder.js';
 const mjAstify = SQLParser.Astify.bind(SQLParser);
